chore: address feedback
- remove constructor
- add owner to Identity struct and not roles
- remove generateRole
- members now have identityId as the role
- emit owner on IdentityCreated

+++ contracts/core/Allo.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 pragma solidity 0.8.19;
 
-import { Metadata } from "./libraries/Metadata.sol";
+import {Metadata} from "./libraries/Metadata.sol";
 import "@openzeppelin-upgradeable/proxy/utils/Initializable.sol";
 
 contract Allo is Initializable {
@@ -21,9 +21,7 @@ contract Allo is Initializable {
 
     // Public getter on the pools mapping
     /// @notice calls out to the registry to get the identity metadata
-    function getPoolInfo(
-        uint256 _poolId
-    ) external view returns (PoolData memory, string memory) {
+    function getPoolInfo(uint256 _poolId) external view returns (PoolData memory, string memory) {
         // Implement the function here
     }
 
@@ -32,9 +30,7 @@ contract Allo is Initializable {
     // creates pool locally, transfers pool amount to distribution strategy => returns poolId
     // takes fee from user
     // validates that the owner is actually allowed to use the identity
-    function createPool(
-        PoolData memory /*_poolData*/
-    ) external pure returns (uint) {
+    function createPool(PoolData memory /*_poolData*/ ) external pure returns (uint256) {
         uint32 _poolId = 0;
 
         // todo: return the poolId? what do we want to return here?
@@ -43,47 +39,41 @@ contract Allo is Initializable {
 
     // passes _data & msg.sender through to the allocation strategy for that pool
     // it should return a uint that represents the application
-    function applyToPool(
-        uint _poolId,
-        bytes memory _data
-    ) external payable returns (uint) {
+    function applyToPool(uint256 _poolId, bytes memory _data) external payable returns (uint256) {
         // Implement the function here
     }
 
     // decode the _data into what's relevant for this strategy
     // perform whatever actions are necessary (token transfers, storage updates, etc)
     // all approvals, checks, etc all happen within internal functions from here?
-    function updateMetadata(
-        uint _poolId,
-        bytes memory _data
-    ) external payable returns (bytes memory) {
+    function updateMetadata(uint256 _poolId, bytes memory _data) external payable returns (bytes memory) {
         // Implement the function here
     }
 
     // transfers _poolAmt from msg.sender to the pool, and takes a fee
-    function fundPool(uint _poolId, uint _poolAmt) external payable {
+    function fundPool(uint256 _poolId, uint256 _poolAmt) external payable {
         // Implement the function here
     }
 
     // passes _data & msg.sender through to the allocation strategy for that pool
-    function allocate(uint _poolId, bytes memory _data) external payable {
+    function allocate(uint256 _poolId, bytes memory _data) external payable {
         // Implement the function here
     }
 
     // calls voting.generatePayouts() and then uses return data for payout.activatePayouts()
     // permissionless for anyone to call, checks happen within the strategies
     // check to make sure they haven't skrited around fee
-    function finalize(uint _poolId) external {
+    function finalize(uint256 _poolId) external {
         // Implement the function here
     }
 
     // call to payoutStrategy.distribute()
-    function distribute(uint _poolId, bytes memory _data) external {
+    function distribute(uint256 _poolId, bytes memory _data) external {
         // Implement the function here
     }
 
     // call to distributionStrategy.close()
-    function closePool(uint _poolId) external {
+    function closePool(uint256 _poolId) external {
         // Implement the function here
     }
 }
